// frontend/src/teamMgt/types.ts
export interface Person {
  id: string;
  name: string;
  picture: string | null; // base64 data URL
}

export interface Team {
  id: string;
  name: string;
  memberIds: string[];
}

export interface TeamsFileData {
  people: Person[];
  teams: Team[];
}

export interface TeamMgtHandle {
  getPeople: () => Person[];
  getTeams: () => Team[];
}